refactor(lua): the objects leave veafSpawnEffects, so its name is true again
`veafSpawnEffects.lua` was the leftover quarter of the May 2026 split of
`veafSpawn.lua`, named after the loudest thing in it. Five of its nine
functions created, moved or removed objects — cargo, the CTLD logistic
unit, `-spawn static`, teleport, destroy — and four were effects.

Those five move to `veafSpawnObjects.lua`. `veafSpawnEffects.lua` keeps
bomb, smoke, signal flare and illumination flare, and nothing else.

No function body is touched: both files, stripped of comments and blank
lines, are identical to the original, 442 lines either way. The 45 Lua
suites pass unchanged, which is this lot's whole acceptance criterion.

`veafSpawnObjects.lua` loads *before* `veafSpawnEffects.lua` in the proxy
and in `LUA_BUNDLE_SCRIPTS`, and that is not cosmetic: command handlers
register at load time into an ordered list and `executeCommand` dispatches
first-match-wins, so the file order is the handler order. This order
reproduces the previous one exactly.

The one step in adding a Lua sub-module that fails in silence — the
`dofile` in the proxy, which yields a nil function at runtime instead of
an error at load — gets a guard: `test_lua_bundle_manifest` now enumerates
`veafSpawn*.lua` from disk and refuses a file the proxy does not load, a
file it loads that no longer exists, and an order diverging from the
bundle. Shown red by deleting the new `dofile` line (2 failed, 5 passed),
green once restored.

Also updates the five comments that named the module for code which is now
in `veafSpawnObjects`. Two of them live in `veafDcsSpawner.lua` and
`test_veafDcsSpawner.lua`, which another lot is editing in parallel.
