+++ .backlog/README.md
@@ -84,7 +84,7 @@ lots are compacted into `.backlog/archive/<LOT-ID>.md`. Sequencing lives in
 | [REFACTOR-SKYNET-WITHOUT-MIST](REFACTOR-SKYNET-WITHOUT-MIST/PRD.md) — **cut Skynet's 42 calls to MiST, in a fork we own.** The last community script that forced the MiST injection. Both repositories upstream are dormant — walder since December 2023, the Regroupement-Patrouille fork since September 2025, with a VEAF pull request there unanswered for 7 days — so VEAF forked the fork. A 258-line compatibility module inside Skynet, with no dependency on VEAF, keeps it a drop-in script and leaves an upstream contribution possible. Artefact verified identical function for function: 282 functions, 140 keys, 17 classes. | ✅ |
 | [FIX-CLONE-KEEPS-ITS-SOURCE-NAME](FIX-CLONE-KEEPS-ITS-SOURCE-NAME/PRD.md) — **every clone of a group is called the same thing.** MiST invented a fresh name for a clone whose name was already taken; the port kept the name unless none was supplied at all. Measured: three clones of `Arco` reach `coalition.addGroup` as three groups named `Arco`, unit names included. Both callers — `veafAirWaves` and `veafQraCore` — clone without a name and then track their groups *by* that name. `isNameTaken` was written for exactly this at ticket 04 and has no caller. | ✅ |
 | [CHORE-ONE-TERRAIN-CHECK](CHORE-ONE-TERRAIN-CHECK/PRD.md) — **six places ask DCS the same question about the ground.** `checkPositionForUnit` holds the real rule (naval wants water, ground wants anything else, air wants 10 m), while `acceptableGroundPoint`, the CSAR survivor placement, `findPointInZone` and `veafSanctuary` each re-derive their own, and the MiST port added a sixth. Two findings: `findPointInZone` already had MiST's ship/LAND-ROAD-RUNWAY rule written out inline, and **`veafSanctuary` compares raw numbers** (`== 2 or == 3`) instead of the surface constants — right today, silently wrong the day they move. **After DROP-MIST 07 and 08.** | ⬜ |
-| [CHORE-RENAME-SPAWN-EFFECTS](CHORE-RENAME-SPAWN-EFFECTS/PRD.md) — **`veafSpawnEffects` is not about effects.** Born 2026-05-21 as the residue of the four-way split of `veafSpawn.lua` (4528 lines) and named after the loudest quarter of what it holds: five of its nine functions create, move or remove objects (cargo, logistic, `-spawn static`, teleport, destroy) and only four are effects. The name already produced one wrong sentence in DROP-MIST ticket 07. Split into `veafSpawnObjects.lua` + a truthful `veafSpawnEffects.lua`; the only unguarded step is the `dofile` in the proxy. **After DROP-MIST 07.** | ⬜ |
+| [CHORE-RENAME-SPAWN-EFFECTS](CHORE-RENAME-SPAWN-EFFECTS/PRD.md) — **`veafSpawnEffects` is not about effects.** Born 2026-05-21 as the residue of the four-way split of `veafSpawn.lua` (4528 lines) and named after the loudest quarter of what it holds: five of its nine functions create, move or remove objects (cargo, logistic, `-spawn static`, teleport, destroy) and only four are effects. The name already produced one wrong sentence in DROP-MIST ticket 07. Split into `veafSpawnObjects.lua` + a truthful `veafSpawnEffects.lua`; the only unguarded step is the `dofile` in the proxy. **Done 2026-09-01**: 376 lines of objects, 188 of effects, proven a pure move — both files stripped of comments and blanks are byte-identical to the original, 442 lines either way. Two findings the PRD did not carry: the file order **is** the command-handler order (`commandHandlers` is an ordered list, dispatch is first-match-wins), so objects load before effects to reproduce it exactly; and the `dofile` it called the one silent risk is in fact already caught by `test_veafSpawn.lua`, which asserts on constants that moved. Guarded anyway, for the next sub-module: three tests enumerate `veafSpawn*.lua` from disk and refuse one the proxy does not load, one it loads and no longer exists, and an order that diverges from the bundle. | ✅ |
 | [FIX-WAVE-OFFSET-AXES](FIX-WAVE-OFFSET-AXES/PRD.md) — **`[latDelta,lonDelta]` moves east and south.** Found 2026-09-01 on the line above the `FIX-AIRWAVES-COMMAND-EASTING` defect, and deliberately left there: the first bracket number is added to the **easting** and the second **subtracted** from the northing, so the two parameters of a public setter — `setRespawnDefaultOffset(defaultOffsetLatitude, defaultOffsetLongitude)`, `@param` tags and all — both mean something other than their names. Consistent across four call sites in two modules, so it was decided once and copied. The documented examples are wrong too, and one of them is wrong under *either* reading, so they do not reveal the intent. **Needs a decision before it starts:** fix the code (the names become true, but every mission with a non-zero offset moves, and anyone who compensated by eye gets it applied twice) or fix the names (nothing moves, but `[east,south]` reads backwards and the setter is public API). Reco: fix the code, migration stated out loud — David's call, they are his missions **Decided 2026-09-01: the code is wrong** — latitude on the northing, longitude on the easting, both added, and the migration announced in the changelog. The documented example `[-3000,0]` annotated "3 km south" is correct only under the corrected reading, which is what settled the intent rather than a preference. Guarded in two layers: eight Lua cases for the direction per axis, and a source test refusing the wrong pairing anywhere — the DCS-group branch reaches its offset only when DCS fails to return a group's units, and reverting that one site leaves all 45 Lua suites green. | ✅ |
 | [FIX-AIRWAVES-COMMAND-EASTING](FIX-AIRWAVES-COMMAND-EASTING/PRD.md) — **a command-driven air wave spawns with a nil easting.** `veafAirWaves.lua:1012` hands `veafInterpreter.execute` the vec2 that `getRandomPointInCircle` answers, while a command expects a vec3 whose `z` is the easting — `veafSpawnGround` reads `spawnPosition.z`. Hidden because the MiST stub in `dcs_mocks` answered a vec3 where MiST answers a vec2, so the test was asserting the mock. Found while porting ticket 06 of DROP-MIST, 2026-08-28. **Fixed 2026-09-01, in two places rather than one:** enumerating the three callers of `veafInterpreter.execute` instead of trusting the PRD found `veafQraCore.lua:994` carrying the identical branch — `veafCombatZone` builds its vec3 by hand and was fine. The PRD's own supporting argument had gone stale meanwhile: the converting "sibling twenty lines down" no longer exists, the `VeafGroupSpawn` chain having absorbed it. Tests tell *correct*, *zero* and *absent* apart and are proven to fail two different ways. Remaining: the in-game look, ticket 02. | 🧑 |
 | [FIX-PLACEMENT-MOVES-ON-CLEAR-GROUND](FIX-PLACEMENT-MOVES-ON-CLEAR-GROUND/PRD.md) — **a FARP escort is moved even when the spot asked for is free.** `findClearBearing` consults the scenery cloud before testing the requested bearing at all, and the wanted spot is never one of the cloud's candidates — so the escort moves whenever the cloud answers. Measured 2026-08-28 on open ground with nothing within a kilometre: bearing 0 requested, 25 used at 1.12×. Testing the requested bearing first is *not* the fix (the occupancy probe cannot see forests); the tenable route is the `gap` tier 1 already computes. **Blocks `FIX-PLACEMENT-IGNORES-SCENERY` ticket 04**, whose definition of done requires proving nothing moves. Fixed and tested 2026-09-01 — a candidate within `PLACEMENT_CLEARANCE` of the wanted spot now proves it, and the wanted spot is kept; only the in-game proof remains | 🧑 |
