+++ test/lua/dcs_mocks.lua
@@ -865,7 +865,7 @@ Controller = {
 -- ctld  (minimal stub — only the API surface used by veafSpawn sub-modules)
 --
 -- Mixed v1 / v2 on purpose, for the length of the CTLD 2 migration: the v1 globals
--- below are still what veafGrass / veafSpawnGround / veafSpawnEffects poke, and they
+-- below are still what veafGrass / veafSpawnGround / veafSpawnObjects poke, and they
 -- go when FEAT-CTLD2-INTEGRATION ticket 05 ports those bridges to the v2 managers.
 -- `utils.log` and `initialize` are the v2 surface veaf.lua drives today.
 -- ---------------------------------------------------------------------------
